backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.12.0
aiohttp==3.9.3
oss2==2.18.4
Pillow==10.2.0
authlib==1.3.0
httpx==0.26.0
SQLAlchemy==2.0.25
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

